My-Blog-ServeOpen/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.1.5
django-cors-headers==3.14.0
django-paginator==0.2a3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.7
mysqlclient==2.1.1
Pillow==9.0.1
PyJWT==2.8.0
pytz==2024.1
qiniu==7.10.0
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
